move BuildTaskServiceEndpoint Token Provider Factory
